stop-place-tile * Move RouteCheckbox and TransportModeSwitch into their own directory
